+++ packages/flutter_chat_ui/lib/src/utils/typedefs.dart
@@ -3,7 +3,8 @@ import 'dart:ui';
 import 'package:flutter_chat_core/flutter_chat_core.dart';
 
 typedef ResolveUserCallback = Future<User?> Function(String id);
-typedef OnMessageTapCallback = void Function(Message message);
+typedef OnMessageTapCallback =
+    void Function(Message message, {required int index});
 typedef OnMessageSendCallback = void Function(String text);
 typedef OnAttachmentTapCallback = VoidCallback;
 typedef PaginationCallback = Future<void> Function();